+++ common/src/main/java/com/wynntils/functions/MinecraftFunctions.java
@@ -31,10 +31,10 @@ public class MinecraftFunctions {
         }
     }
 
-    public static class DirFunction extends Function<Float> {
+    public static class DirFunction extends Function<Double> {
         @Override
-        public Float getValue(FunctionArguments arguments) {
-            return McUtils.player().getYRot();
+        public Double getValue(FunctionArguments arguments) {
+            return (double) McUtils.player().getYRot();
         }
     }
